+++ src/main/java/org/example/fanzip/global/config/ServletConfig.java
@@ -1,4 +1,4 @@
-package org.example.fanzip.config;
+package org.example.fanzip.global.config;
 
 import lombok.RequiredArgsConstructor;
 import org.example.fanzip.auth.jwt.JwtInterceptor;
