+++ src/main/java/org/example/fanzip/global/controller/TestController.java
@@ -1,4 +1,4 @@
-package org.example.fanzip.controller;
+package org.example.fanzip.global.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
