+++ src/test/java/org/example/fanzip/user/service/UserServiceTest.java
@@ -1,13 +1,16 @@
 package org.example.fanzip.user.service;
 
 import lombok.extern.slf4j.Slf4j;
-import org.example.fanzip.config.RootConfig;
+import org.example.fanzip.global.config.RootConfig;
+//import org.example.fanzip.user.dto.AdditionalInfoDTO;
 import org.example.fanzip.user.dto.UserDTO;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.util.Date;
 
 import static org.junit.jupiter.api.Assertions.*;
